Fix some review mentions
